feat: Enable stega in Vercel preview deployments
This commit enables stega if it's a Vercel preview deployment by setting the `enabled` flag to true. It also updates the revalidation time for queries based on whether or not the CDN is being used. Additionally, it sets up VisualEditing to always use stega and allows only same-origin Studios to connect.