feat: Import and use `urlForImage` function in `[slug]/page.tsx`
This commit imports the `urlForImage` function from `@/lib/sanity.image` and uses it to generate a smaller image URL for the painting preview. The generated URL is now used as the image source in the JSON-LD metadata.

The commit also updates the revalidation time for the landing page to 1 week, and removes commented out code related to generating metadata.

In addition, this commit updates the revalidation times for other pages: paintings page (3 hours), projects page (3 hours), videos page (3 hours).

Lastly, this commit fixes a typo in `loadQuery.ts`, changing `"painting/${slug}"` to `"painting:${slug}"`.